Create TO-DO LIST CODE
A to-do list is a list of items that need to be completed. The items on the list can range from simple activities like replying to an email,to more complex tasks like creating project briefs.A TO-DO LIST contains all the tasks that need to be done and can be ticked off accordingly.